conda-requirements.txt
python=3.9
numpy=1.23.0
pyoptsparse
matplotlib
pandas
scienceplots
pyright
black
mpi4py
petsc
petsc4py
scikits.odes
